Retrieve transfer user from jwt token
