+++ cloud-pipeline-common/model/src/main/java/com/epam/pipeline/dto/dts/transfer/TaskCreationDTO.java
@@ -28,5 +28,4 @@ public class TaskCreationDTO {
     private StorageItemWithCredentialsDTO source;
     private StorageItemWithCredentialsDTO destination;
     private List<String> included = new ArrayList<>();
-    private String user;
 }
