+++ data-transfer-service/src/main/java/com/epam/pipeline/dts/security/service/impl/SecurityServiceImpl.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2017-2020 EPAM Systems, Inc. (https://www.epam.com/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.epam.pipeline.dts.security.service.impl;
+
+import com.epam.pipeline.dts.security.exception.SecurityServiceException;
+import com.epam.pipeline.dts.security.service.SecurityService;
+import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@RequiredArgsConstructor
+@Service
+public class SecurityServiceImpl implements SecurityService {
+    
+    @Override
+    public String getAuthorizedUser() {
+        return Optional.of(SecurityContextHolder.getContext())
+                .map(SecurityContext::getAuthentication)
+                .map(Authentication::getName)
+                .filter(StringUtils::isNotBlank)
+                .orElseThrow(() -> new SecurityServiceException("Authorized user name is missing."));
+    }
+}
