+++ data-transfer-service/src/main/java/com/epam/pipeline/dts/transfer/rest/controller/TransferController.java
@@ -70,8 +70,7 @@ public class TransferController extends AbstractRestController {
         TransferTask task = transferService.runTransferTask(
                 storageItemMapper.dtoToModel(taskCreationDTO.getSource()),
                 storageItemMapper.dtoToModel(taskCreationDTO.getDestination()),
-                taskCreationDTO.getIncluded(),
-                taskCreationDTO.getUser());
+                taskCreationDTO.getIncluded());
         return Result.success(taskMapper.modelToDto(task));
     }
