+++ data-transfer-service/src/main/resources/application.properties
@@ -62,3 +62,5 @@ dts.service.name=DTS
 dts.pipeline.read.timeout=30
 #in sec
 dts.pipeline.connect.timeout=10
+
+dts.username.transformation=NONE
